updated services, split AuroraConfigService into two services, fixed tests
