+++ src/main/kotlin/no/skatteetaten/aurora/gobo/resolvers/scan/ScanQueryResolver.kt
@@ -1,11 +1,11 @@
 package no.skatteetaten.aurora.gobo.resolvers.scan
 
 import com.coxautodev.graphql.tools.GraphQLQueryResolver
-import no.skatteetaten.aurora.gobo.integration.unclematt.ProbeService
+import no.skatteetaten.aurora.gobo.integration.unclematt.ProbeServiceBlocking
 import org.springframework.stereotype.Component
 
 @Component
-class ScanQueryResolver(val scanService: ProbeService) : GraphQLQueryResolver {
+class ScanQueryResolver(val scanService: ProbeServiceBlocking) : GraphQLQueryResolver {
 
     fun scan(host: String, port: Int = 80): Scan {
         return Scan.fromProbeResultList(scanService.probeFirewall(host, port))
