+++ src/test/kotlin/no/skatteetaten/aurora/gobo/service/ApplicationUpgradeServiceTest.kt
@@ -9,9 +9,11 @@ import no.skatteetaten.aurora.gobo.ApplicationDeploymentDetailsBuilder
 import no.skatteetaten.aurora.gobo.AuroraConfigFileBuilder
 import no.skatteetaten.aurora.gobo.integration.SourceSystemException
 import no.skatteetaten.aurora.gobo.integration.boober.AuroraConfigService
+import no.skatteetaten.aurora.gobo.integration.boober.BooberWebClient
 import no.skatteetaten.aurora.gobo.integration.boober.Response
 import no.skatteetaten.aurora.gobo.integration.execute
 import no.skatteetaten.aurora.gobo.integration.mokey.ApplicationService
+import no.skatteetaten.aurora.gobo.integration.mokey.ApplicationServiceBlocking
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import okhttp3.mockwebserver.SocketPolicy
@@ -25,9 +27,8 @@ class ApplicationUpgradeServiceTest {
 
     private val config = ApplicationConfig("${url}mokey", "${url}unclematt", 30000, 30000)
 
-    private val applicationService = ApplicationService(config.webClientMokey())
-    private val auroraConfigService = AuroraConfigService("${url}boober", config.webClientBoober())
-
+    private val auroraConfigService = AuroraConfigService(BooberWebClient("${url}boober", config.webClientBoober()))
+    private val applicationService = ApplicationServiceBlocking(ApplicationService(config.webClientMokey()))
     private val upgradeService = ApplicationUpgradeService(applicationService, auroraConfigService)
 
     @Test
@@ -39,7 +40,7 @@ class ApplicationUpgradeServiceTest {
             redeployResponse(),
             enqueueRefreshResponse()
         ) {
-            upgradeService.upgrade("applicationDeploymentId", "version", "token")
+            upgradeService.upgrade("token", "applicationDeploymentId", "veresion")
         }
 
         assert(requests[0].path).isEqualTo("/mokey/api/auth/applicationdeploymentdetails/applicationDeploymentId")
